third_party/WebKit/LayoutTests/http/tests/loading/preload-video-poster.html

<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=100"></script>
</head>
<body>
<script>
    shouldBeTrue("internals.isPreloaded('resources/base-image1.png');");
    shouldBeTrue("internals.isPreloaded('resources/base-image2.png');");
    shouldBeFalse("internals.isPreloaded('resources/base-image3.png');");
</script>
<div>This test passes if the image (defined in the src attribute) is preloaded.</div>
<video poster="resources/base-image1.png"></video>
<video poster="resources/base-image2.png" poster="resources/base-image3.png"></video>
</body>
</html>
